Added NLP data set and ability to pull html DUA forms from SciAuthZ
Added the NLP data set to the fixtures file.

Updated the SciAuthZ request functions to expect a field containing the DUA form (an html string). If we get back a form, we display that html which will contain input fields for the user to fill in. Once the user hits submit, the full DUA including the user's inputs will be collected and sent back to SciAuthZ to complete the request. Some jQuery code is used to replace the input fields with their values, allowing us to grab the completed DUA form as one long string.

Removed jwt package which is not needed and causes issues with the PyJWT package.

Note that in development you may need to add verify=False flags to all request  calls because the hostname on the certificate is not localhost.