TC-50: Implemented registration sign up for new users
If a user hit hypatio for the first time, a user gets created for them in scireg automatically, but not the registration form. The profile page will not prompt them to complete their registration and then it will send that data to scireg to be created as a new Registration object. Also added functionality to make the first and last name readonly fields if the user already has a registration.