TC-38: Adds a screen for project authorization and permission granting
This commit adds a Manage Contests screen where a user with MANAGE permissions on a given item (for now just "N2C2") can view all the users who have requested authorization to a project. The user can then grant the user authorization to the project and VIEW permissions at the click of a button. Calls will be made to SciAuthZ to do this.

Also downgrades from bootstrap4 to 3 for better compatibility with existing Hypatio functionality.