TC-79: n2c2 real description and several UI improvements
Includes a migration to make the description field a TextField (so it's larger). Adds to the fixture file the full description for the n2c2 project and changes its name to include 'track 1'.  Moves the login_url logic to a template filter since this is used in several views and templates. Adjusts the project_details view to show a login/register message if user not logged in (instead of showing forms). Adjusts CSS to only show the red x/green checkmark on panels related to project registration. Fixes home link to actually point to hypatio home page. Adjusts the page header to have the same width/alignment as the nav bar and panels below it.