TC-117: Brings profile registration form into project page
Moved the profile registration form to a separate html file so both the project and profile pages can include it. Added a separate view to handle form processing so that logic doesn't need to duplicated in the project and profile views. I also unbinded the data to the form in the project view, so the user is not overwhelmed with a green/red validation highlighting.