TC-179: Accepting and rejecting forms
Signed forms now have statuses of either accepted, rejected, or pending. On the challenge management screen, they are highlighted with different colors to indicate their status. Clicking on a signed form brings it up in a new window, with buttons to change its status. When rejecting a form, the user will be prompted to explain why and that message will be sent to the affected person by email. If the team is already in an Active state when this happens, the team will be dropped to a Ready state so it can be reviewed again, and this will remove the team members' VIEW permissions. Also, the button to activate a team will be disabled if any team member has not yet filled in a form or does not have all their forms approved.