TC-186: Permission checks and download capability
Checks if a user has the proper permissions to a project before uploading their submission. Increases the expires value for S3 uploads to support clients with a small time difference compared to the server. Separates S3 file storage for project files from participant submissions into different buckets. Adds a table to the manage team screen for admins to be able to view participant submissions. Adds functionality to download those participant submissions if the user is a manager of the project.