TC-202: Adds submission form and stores answer in ParticipantSubmission object
Projects can now have submission forms that are dynamically pulled from an html file and displayed for participants to fill out when uploading their file. The response to the form is serialized into a json and then stored in the ParticipantSubmission object.
This commit also fixes a small bug with new users where profile_registration_info['count'] is not an available key from scireg.