TC-206 / TC-205: Download a team's submissions
Adds a download button on the team management page for an admin to be able to download all of a team's submissions in one zip file. The zip file contains additional zip files for each individual submission. Those zip files hold the originally uploaded file along with a json file holding the results of the submission form completed by the user.