TC-208: Forms now pulled from static folder
Agreement forms and submission forms were previously pulled from the media folder in order to enable future uploads by users managing contests. However, this complicates the deployment process as a fixture would be needed to load those files. So I moved the files into the static folder and renamed the files with real names so they're easier to track.

Also added a visible error when someone unauthorized tries to delete a submission.