HYP-71: Refactoring submissions into ChallengeTasks
N2C2 Track 2 required more than one type of submission: "tasks". This commit renames the ParticipantSubmission model to ChallengeTaskSubmission and adds a ChallengeTask model, of which a DataProject can have multiple. This then required changes to viewing, uploading, deleting, and downloading submissions, as well as a custom migration to preserve N2C2 Track 1 submissions that did not have a ChallengeTask concept at the time. Also includes a number of UI changes to the challenge participate screen.