HYP-80: Fixes a migration file bug regarding an unknown column
For some reason, a column name was not updated in an intermediary table of a manytomanyfield when the related model was renamed. I couldn't find out why Django fails to update the column name when the model was renamed, so instead I am using raw SQL in a migration file to accomplish this.