HYP-67: Basic editing of existing hostedfiles on new manage page
Enables basic editing of an existing hosted file. More work to be done on showing errors, changing the file itself, and picking a HostedFileSet. Also, I removed all the DataGate stuff because they aren't used and can be optimized in a cleaner way later. For now, HostedFiles will only be from S3.