HYP-36/HYP-35: Add screen to view projects you manage and other refactoring
Adds a new view where users who have any kind of MANAGE permission will be able to see a list of projects they manage and then go to their respective management screens. Also refactors a lot of repetitive code to check if someone is a manager and then display the button in the nav bar to go to the manage projects page. Also greatly reduces the length of code in the list_data_projects and list_data_challenges views which had a lot of code no longer needed.