HYP-100: Complete remaining DataProject view refactoring
This commit adds the ability for Participant records to be created for projects without teams so that a single person may submit (and delete) solutions to tasks. This also adds the HTML submission form for the new n2c2 data upload project.

Also fixes emails which were not sending in a local environment.