HYP-102: Refactor submission downloads and enable single submission download
Previously, could only download all submissions by a team at a time. Refactored the code to allow a user to download a single submission too. Replaced the TeamSubmissionDownload model with a ChallengeTaskSubmissionDownload model to enable logging of single submission downloads. Added a table to the project management page for managers to see all submissions and be able to download them.